Migrated the FCM impl to the new HTTPClient API
