+++ .gitignore
@@ -1,2 +1,4 @@
 testdata/integration_*
 .vscode/*
+*~
+
