+++ messaging/messaging_test.go
@@ -22,7 +22,6 @@ import (
 	"net/http"
 	"net/http/httptest"
 	"reflect"
-	"strings"
 	"testing"
 	"time"
 
@@ -832,41 +831,6 @@ var invalidMessages = []struct {
 	},
 }
 
-var invalidTopicMgtArgs = []struct {
-	name   string
-	tokens []string
-	topic  string
-	want   string
-}{
-	{
-		name: "NoTokensAndTopic",
-		want: "no tokens specified",
-	},
-	{
-		name:   "NoTopic",
-		tokens: []string{"token1"},
-		want:   "topic name not specified",
-	},
-	{
-		name:   "InvalidTopicName",
-		tokens: []string{"token1"},
-		topic:  "foo*bar",
-		want:   "invalid topic name: \"foo*bar\"",
-	},
-	{
-		name:   "TooManyTokens",
-		tokens: strings.Split("a"+strings.Repeat(",a", 1000), ","),
-		topic:  "topic",
-		want:   "tokens list must not contain more than 1000 items",
-	},
-	{
-		name:   "EmptyToken",
-		tokens: []string{"foo", ""},
-		topic:  "topic",
-		want:   "tokens list must not contain empty strings",
-	},
-}
-
 func TestNoProjectID(t *testing.T) {
 	client, err := NewClient(context.Background(), &internal.MessagingConfig{})
 	if client != nil || err == nil {
@@ -1021,7 +985,7 @@ func TestSendError(t *testing.T) {
 		t.Fatal(err)
 	}
 	client.fcmEndpoint = ts.URL
-	client.client.RetryConfig = nil
+	client.fcmClient.httpClient.RetryConfig = nil
 
 	for _, tc := range httpErrors {
 		resp = tc.resp
@@ -1045,150 +1009,6 @@ func TestInvalidMessage(t *testing.T) {
 				t.Errorf("Send(%s) = (%q, %v); want = (%q, %q)", tc.name, name, err, "", tc.want)
 			}
 		})
-	}
-}
-
-func TestSubscribe(t *testing.T) {
-	var tr *http.Request
-	var b []byte
-	ts := httptest.NewServer(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
-		tr = r
-		b, _ = ioutil.ReadAll(r.Body)
-		w.Header().Set("Content-Type", "application/json")
-		w.Write([]byte("{\"results\": [{}, {\"error\": \"error_reason\"}]}"))
-	}))
-	defer ts.Close()
-
-	ctx := context.Background()
-	client, err := NewClient(ctx, testMessagingConfig)
-	if err != nil {
-		t.Fatal(err)
-	}
-	client.iidEndpoint = ts.URL
-
-	resp, err := client.SubscribeToTopic(ctx, []string{"id1", "id2"}, "test-topic")
-	if err != nil {
-		t.Fatal(err)
-	}
-	checkIIDRequest(t, b, tr, iidSubscribe)
-	checkTopicMgtResponse(t, resp)
-}
-
-func TestInvalidSubscribe(t *testing.T) {
-	ctx := context.Background()
-	client, err := NewClient(ctx, testMessagingConfig)
-	if err != nil {
-		t.Fatal(err)
-	}
-	for _, tc := range invalidTopicMgtArgs {
-		t.Run(tc.name, func(t *testing.T) {
-			resp, err := client.SubscribeToTopic(ctx, tc.tokens, tc.topic)
-			if err == nil || err.Error() != tc.want {
-				t.Errorf(
-					"SubscribeToTopic(%s) = (%#v, %v); want = (nil, %q)", tc.name, resp, err, tc.want)
-			}
-		})
-	}
-}
-
-func TestUnsubscribe(t *testing.T) {
-	var tr *http.Request
-	var b []byte
-	ts := httptest.NewServer(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
-		tr = r
-		b, _ = ioutil.ReadAll(r.Body)
-		w.Header().Set("Content-Type", "application/json")
-		w.Write([]byte("{\"results\": [{}, {\"error\": \"error_reason\"}]}"))
-	}))
-	defer ts.Close()
-
-	ctx := context.Background()
-	client, err := NewClient(ctx, testMessagingConfig)
-	if err != nil {
-		t.Fatal(err)
-	}
-	client.iidEndpoint = ts.URL
-
-	resp, err := client.UnsubscribeFromTopic(ctx, []string{"id1", "id2"}, "test-topic")
-	if err != nil {
-		t.Fatal(err)
-	}
-	checkIIDRequest(t, b, tr, iidUnsubscribe)
-	checkTopicMgtResponse(t, resp)
-}
-
-func TestInvalidUnsubscribe(t *testing.T) {
-	ctx := context.Background()
-	client, err := NewClient(ctx, testMessagingConfig)
-	if err != nil {
-		t.Fatal(err)
-	}
-	for _, tc := range invalidTopicMgtArgs {
-		t.Run(tc.name, func(t *testing.T) {
-			resp, err := client.UnsubscribeFromTopic(ctx, tc.tokens, tc.topic)
-			if err == nil || err.Error() != tc.want {
-				t.Errorf(
-					"UnsubscribeFromTopic(%s) = (%#v, %v); want = (nil, %q)", tc.name, resp, err, tc.want)
-			}
-		})
-	}
-}
-
-func TestTopicManagementError(t *testing.T) {
-	var resp string
-	ts := httptest.NewServer(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
-		w.WriteHeader(http.StatusInternalServerError)
-		w.Header().Set("Content-Type", "application/json")
-		w.Write([]byte(resp))
-	}))
-	defer ts.Close()
-
-	ctx := context.Background()
-	client, err := NewClient(ctx, testMessagingConfig)
-	if err != nil {
-		t.Fatal(err)
-	}
-	client.iidEndpoint = ts.URL
-	client.client.RetryConfig = nil
-
-	cases := []struct {
-		resp, want string
-		check      func(error) bool
-	}{
-		{
-			resp:  "{}",
-			want:  "http error status: 500; reason: client encountered an unknown error; response: {}",
-			check: IsUnknown,
-		},
-		{
-			resp:  "{\"error\": \"INVALID_ARGUMENT\"}",
-			want:  "http error status: 500; reason: request contains an invalid argument; code: invalid-argument",
-			check: IsInvalidArgument,
-		},
-		{
-			resp:  "{\"error\": \"TOO_MANY_TOPICS\"}",
-			want:  "http error status: 500; reason: client exceeded the number of allowed topics; code: too-many-topics",
-			check: IsTooManyTopics,
-		},
-		{
-			resp:  "not json",
-			want:  "http error status: 500; reason: client encountered an unknown error; response: not json",
-			check: IsUnknown,
-		},
-	}
-	for _, tc := range cases {
-		resp = tc.resp
-		tmr, err := client.SubscribeToTopic(ctx, []string{"id1"}, "topic")
-		if err == nil || err.Error() != tc.want || !tc.check(err) {
-			t.Errorf("SubscribeToTopic() = (%#v, %v); want = (nil, %q)", tmr, err, tc.want)
-		}
-	}
-	for _, tc := range cases {
-		resp = tc.resp
-		tmr, err := client.UnsubscribeFromTopic(ctx, []string{"id1"}, "topic")
-		if err == nil || err.Error() != tc.want {
-			t.Errorf("UnsubscribeFromTopic() = (%#v, %v); want = (nil, %q)", tmr, err, tc.want)
-		}
 	}
 }
 
@@ -1226,50 +1046,6 @@ func checkFCMRequest(t *testing.T, b []byte, tr *http.Request, want map[string]i
 	clientVersion := "fire-admin-go/" + testMessagingConfig.Version
 	if h := tr.Header.Get("X-FIREBASE-CLIENT"); h != clientVersion {
 		t.Errorf("X-FIREBASE-CLIENT = %q; want = %q", h, clientVersion)
-	}
-}
-
-func checkIIDRequest(t *testing.T, b []byte, tr *http.Request, op string) {
-	var parsed map[string]interface{}
-	if err := json.Unmarshal(b, &parsed); err != nil {
-		t.Fatal(err)
-	}
-	want := map[string]interface{}{
-		"to":                  "/topics/test-topic",
-		"registration_tokens": []interface{}{"id1", "id2"},
-	}
-	if !reflect.DeepEqual(parsed, want) {
-		t.Errorf("Body = %#v; want = %#v", parsed, want)
-	}
-
-	if tr.Method != http.MethodPost {
-		t.Errorf("Method = %q; want = %q", tr.Method, http.MethodPost)
-	}
-	wantOp := "/" + op
-	if tr.URL.Path != wantOp {
-		t.Errorf("Path = %q; want = %q", tr.URL.Path, wantOp)
-	}
-	if h := tr.Header.Get("Authorization"); h != "Bearer test-token" {
-		t.Errorf("Authorization = %q; want = %q", h, "Bearer test-token")
-	}
-}
-
-func checkTopicMgtResponse(t *testing.T, resp *TopicManagementResponse) {
-	if resp.SuccessCount != 1 {
-		t.Errorf("SuccessCount = %d; want  = %d", resp.SuccessCount, 1)
-	}
-	if resp.FailureCount != 1 {
-		t.Errorf("FailureCount = %d; want  = %d", resp.FailureCount, 1)
-	}
-	if len(resp.Errors) != 1 {
-		t.Fatalf("Errors = %d; want = %d", len(resp.Errors), 1)
-	}
-	e := resp.Errors[0]
-	if e.Index != 1 {
-		t.Errorf("ErrorInfo.Index = %d; want = %d", e.Index, 1)
-	}
-	if e.Reason != "unknown-error" {
-		t.Errorf("ErrorInfo.Reason = %s; want = %s", e.Reason, "unknown-error")
 	}
 }
